Tracks without album or artist
Now the fulltree for an artist includes a `no_album_tracks` attribute, which
contains the list of tracks for that artist that are not related to any album.
And the `/tracks` resource can retrieve the list of tracks that have no album
(`/tracks?album=0`), no artist (`/tracks?artist=0`), or neither
(`/tracks?album=0&artist=0`).